+++ src/System/IO/Streams/Combinators.hs
@@ -22,9 +22,11 @@ module System.IO.Streams.Combinators
  , map
  , mapM
  , mapM_
+ , mapMaybe
  , contramap
  , contramapM
  , contramapM_
+ , contramapMaybe
 
    -- * Filter
  , filter
@@ -370,6 +372,33 @@ mapM_ f s = makeInputStream $ do
 
 
 ------------------------------------------------------------------------------
+-- | A version of map that discards elements
+--
+-- @mapMaybe f s@ passes all output from @s@ through the function @f@ and
+-- discards elements for which @f s@ evaluates to 'Nothing'.
+--
+-- Example:
+--
+-- @
+-- ghci> Streams.'System.IO.Streams.fromList' [Just 1, None, Just 3] >>=
+--       Streams.'mapMaybe' 'id' >>=
+--       Streams.'System.IO.Streams.toList'
+-- [1,3]
+-- @
+--
+-- /Since: 1.2.1.0/
+mapMaybe :: (a -> Maybe b) -> InputStream a -> IO (InputStream b)
+mapMaybe f src = makeInputStream g
+  where
+    g = do
+      s <- read src
+      case s of
+        Nothing -> return Nothing
+        Just x ->
+          case f x of
+            Nothing -> g
+            y -> return y
+------------------------------------------------------------------------------
 -- | Contravariant counterpart to 'map'.
 --
 -- @contramap f s@ passes all input to @s@ through the function @f@.
@@ -415,6 +444,24 @@ contramapM_ :: (a -> IO b) -> OutputStream a -> IO (OutputStream a)
 contramapM_ f s = makeOutputStream $ \mb -> do
     _ <- maybe (return $! ()) (void . f) mb
     write mb s
+
+
+------------------------------------------------------------------------------
+-- | Contravariant counterpart to 'contramapMaybe'.
+--
+-- @contramap f s@ passes all input to @s@ through the function @f@.
+-- Discards all the elements for which @f@ returns 'Nothing'.
+--
+-- /Since: 1.2.1.0/
+--
+contramapMaybe :: (a -> Maybe b) -> OutputStream b -> IO (OutputStream a)
+contramapMaybe f s = makeOutputStream $ g
+    where
+      g Nothing = write Nothing s
+      g (Just a) =
+        case f a of
+          Nothing -> return ()
+          x -> write x s
 
 
 ------------------------------------------------------------------------------