+++ test/System/IO/Streams/Tests/Combinators.hs
@@ -46,7 +46,9 @@ tests = [ testFilter
         , testContramap
         , testMapM
         , testMapM_
+        , testMapMaybe
         , testContramapM_
+        , testContramapMaybe
         , testSkipToEof
         , testZip
         , testZipWith
@@ -128,6 +130,15 @@ testMapM_ = testCase "combinators/mapM_" $ do
 
 
 ------------------------------------------------------------------------------
+testMapMaybe :: Test
+testMapMaybe = testCase "combinators/mapMaybe" $ do
+    is <- fromList [1,2,3::Int] >>= S.mapMaybe (\x -> if odd x then Just (x * x) else Nothing)
+    l  <- toList is
+
+    assertEqual "mapMaybe" [1,9] l
+
+
+------------------------------------------------------------------------------
 testContramapM_ :: Test
 testContramapM_ = testCase "combinators/contramapM_" $ do
     ref <- newIORef 0
@@ -135,6 +146,15 @@ testContramapM_ = testCase "combinators/contramapM_" $ do
     _   <- outputToList (contramapM_ (modifyIORef ref . (+)) >=> connect is)
 
     readIORef ref >>= assertEqual "contramapM_" 6
+
+
+------------------------------------------------------------------------------
+testContramapMaybe :: Test
+testContramapMaybe = testCase "combinators/contramapMaybe" $ do
+    is  <- fromList [1,2,3::Int]
+    l   <- outputToList (contramapMaybe f >=> connect is)
+    assertEqual "contramapMaybe" [1,9] l
+    where f x = if even x then Nothing else Just $ x * x
 
 
 ------------------------------------------------------------------------------